network: use `qubes-primary-dns` QubesDB entry if present
For a long time the DNS address was the same as default gateway. This is
still the case in R3.x, but using `qubes-gateway` configuration
parameter for it is misleading. It should be up to dom0 to provide DNS
address (whether the value is the same as gateway or not).

Fixes QubesOS/qubes-issues#1817

(cherry picked from commit fb9b3b62c02427b80657056805a9bac7756c98b3)