qubes.Backup: also syncfs() or sync() before *start*
This is for a scenario where $args is e.g. /foo/bak/ and the user
created the bak directory *just* before starting a tiny backup run.
Consider what could happen if we write the backup file, even fsync() it
as well as bak, exit successfully - and then there's a power outage: Now
the backup file is on disk, as is bak (containing an entry referencing
the backup file), but where is the reference *to* bak? POSIX does not
guarantee that foo had ever been persisted after the user created bak as
an entry in it. So bak is possibly an orphan, making the backup file an
orphan too. In the worst case, both are garbage collected collected next
time the filesystem is mounted...