Fix GSettings when XDG_DATA_DIRS is overridden
GSettings requires that glib-2.0 be found in XDG_DATA_DIRS or the
program crashes.  To avoid making an already long line much longer, move
the list of symlinks that need to be made to a separate file.

Tested by (1) creating the symlink manually in a qube and finding that
the problem went away and (2) checking that `make install` worked
properly.