Disable renaming Xen's eth0 to enX0
There are a couple of issues with this renaming:
1. When enabled, the interface name cannot be prediced until it actually
   happens. This breaks waiting for the device to appear in
   qubes-network-uplink.service.
2. Setting SYSTEMD_WANTS on a device that gets renamed seems to not work
   (is the variable bound to the old device name?). This breaks dynamic
   network attach (see 99-qubes-network.rules).

So, disable it completely for Xen devices, at least for now. This may
pose some issues (or rather - rollback fix attempt) for VMs with both
physical devices and Xen netfront device(s), but this is extremely rare
case that nobody complained about before.

Fixes QubesOS/qubes-issues#7284