network: fix uplink setup relation to network.target
qubes-network-uplink@.service was missing a Before=network.target
ordering. Such an ordering is redundant during startup, because these
instances are started through the script in qubes-network-uplink.service
(which already has the ordering), but it's not redundant during
shutdown: The instances must be *stopped after* network consumer units.

Also, network provider units are the ones who are supposed to pull in
the passive network.target with a Wants= dependency. But
qubes-network-uplink@.service and qubes-network-uplink.service didn't.
It was only pulled in coincidentally if another network provider unit
was installed and enabled, such as NetworkManager.service.

Fixes QubesOS/qubes-issues#8005