Replace iptables with nftables
nftables is the modern replacement for iptables.  It has more features
and is actively maintained.  iptables is deprecated and may be removed
from distributions in the future.

Instead of a 1-to-1 translation, update the code to use nftables
features such as sets and maps.  This makes anti-spoofing checks either
O(1) or O(log N) (depending on how sets and maps are implemented) in the
number of downstream network interfaces.  The masquerading path benefits
even more, as it now uses netdev rules to directly forward a packet
to a specific destination while statelessly changing its addresses.

This also removes the legacy iptables support from the firewall daemon.
Since the anti-spoofing rules require nftables, supporting iptables in
the firewall daemon makes no sense anymore.

Furthermore, software flow offload allows NATd traffic to bypass all
layer 3 processing.  This is implemented using flowtables, which use the
ingress hook and therefore come before the prerouting hook.

DNS is now handled entirely in Python, rather than by a combination of
Python and Bash scripts.  This allows much better error checking and
makes the code actually possible to understand.  It will also be much
more extensible in the future.  As just one example, this Python code
could be used by the firewall daemon to automatically update DNS in
response to systemd-resolved events.

Instead of using network interface names, the nftables rules use
interface groups.  This should speed up processing (though no benchmarks
have been done) and avoids problems if interface names change in the
future.  The only exception is the prerouting hooks, which use interface
names to ensure that interface ID reuse does not cause problems.

The new anti-spoofing rules are deny-by-default.  If there is no entry
in the "allowed" nftables map for a given IP protocol, all traffic with
that protocol will be considered spoofed and dropped.  This avoids
having to explicitly disable IPv6 on interfaces with no ipv6 addresses.