Add permanent neighbor entries again
Previously, neighbor discovery used ARP or NDP, both of which have
nonzero attack surface.  Additionally, the strict anti-spoofing rules
used by Qubes OS tend to break NDP, causing IPv6 traffic to be
unreliable.

This solves the second problem by adding permanent entries to Linux's
neighbor cache.  Since the peer MAC address is already known, it is not
necessary to use ARP or NDP to discover it.  This allows NDP to be
blocked outright if it has not been blocked already.