(Re-)enable override-redirect windows in xfce4-notifyd
Upstream decided to not use override-redirect windows for notifications
and use _MOTIF_WM_HINTS to mark those windows as borderless. The reason
was to not show the notifications over the screenlocker. In Qubes, we
have no intention of passing all possible window properties to the real
X server, but also the screenlocker issue doesn't apply as the
gui-daemon takes care to not show those above.

Thanks @rustybird for finding the solution.
Fixes QubesOS/qubes-issues#8181