Set default text/plain editor to something not LibreOffice
Debian does not define default application, so the choice is a bit
random (in practice - lexicographical). For text/plain, it hits
LibreOffice, which is very unfriendly for text/plain. Choose something
that is not libreoffice instead.
Fedora does define gedit as default editor, but xdg-open doesn't handle
mimeapps.list, it only looks for defaults.list file. Add a symlink to
fix this (but not on Ubuntu, which ships the symlink via
desktop-file-utils package).

Re-calculate the choice on package update too, if the currently selected
editor doesn't exist anymore (happens for example when org.gnome.gedit is replaced
with org.gnome.TextEditor).

Fixes QubesOS/qubes-issues#8385