qfile-unpacker should not be an easy root hole
qfile-unpacker is setuid root so that it can call chroot().  Right now,
it will happily change its filesystem UID to a value that is supplied on
the command line, then chroot to a directory supplied on the command
line, and finally create new files, directories, and symbolic links of
the caller's choice.  This is enough to get root; a simple method is to
drop a .conf file under /rw/bind-dirs (creating it if it does not exist)
and wait for VM reboot.  The file will then be sourced as a shell
script with root privileges.

Additionally, setuid() failing results in zenity or kdialog being run
as root, which is not safe.  Neither zenity nor kdialog is safe when
being run with privileges, and it is easy to make them execute arbitrary
code by abusing GTK or Qt's environment variables.  Fortunately,
setuid() failing is not common or easy to trigger.

While Qubes OS does not rely on user -> root to be a security boundary,
it also does not try to sabotage that boundary, at least for users
outside of the "qubes" group (which is considered to be privileged).
Allowing arbitrary users (even "nobody", "guest", or "www") to get root
is generally frowned upon.  Therefore, this problem should be fixed.

To fix this problem, refuse to change to a UID that is not the same as
the real UID of the calling process, and do not invoke a GUI program if
dropping privileges fails.

(cherry picked from commit 25ce49cda633e1a6cf0cce5b4c4d11b980d2de32)