Detect and terminate infinite qrexec launch chains
If a VM is set to use a disposable VM to handle URLs and is also its
also default disposable VM template, attempting to open a URL in that VM
will cause the newly created disposable VM to create yet another
disposable VM and so on.  This is effectively a fork bomb, and will
continue to create disposable VMs until a resource (usually memory)
is exhausted.

To prevent this from happening, check if the VM was spawned via
qubes.OpenURL or qubes.OpenInVM, and if the target of a call to
qvm-open-in-vm is a disposable VM.  If both hold, prompt the user before
creating the disposable VM.  If prompting is not possible, fail.

Tested with a modified Fedora template.

Fixes: QubesOS/qubes-issues#7406