rpm: disable symlink checker on openSUSE for now
/usr/share/qubes/xdg-overrides ships a bunch of symlinks not present
during build. This will require adding proper BuildRequires providing
those files, but figure it out later, especially as some of them are
dynamically generated. The actual use of /usr/share/qubes/xdg-overrides
is optional, and is meant to cover all relevant xdg-related
subdirectories of /usr/share, even if not used on a given system.

QubesOS/qubes-issues#6567