Do not show "Displays" prompt in every VM when new display is connected
Default Xfce setting is to prompt what to do when new display is
connected. It makes sense in dom0 (or GUI domain), but not really in
every VM. Disable it by default. When setting up GUI domain, it will
need to be re-enabled manually (or possibly via salt).

Fixes QubesOS/qubes-issues#8756