Ensure XDG secrets portal is available
Sandboxed applications use this to store secrets.  Without it, Element
cannot access its secrets and resets itself, causing user data to be
lost.  The version of xdg-desktop-portal in Fedora 38 and below would
use any portal implementation it could find.  Since gnome-keyring is
present, this worked fine.  However, xdg-desktop-portal 1.18.2 requires
portal implementations to be specified in a configuration file,
presumably to ensure that the correct portal implementation is
chosen when multiple implementations are installed on a system.

To solve this problem, provide a configuration file that selects
gnome-keyring for org.freedesktop.impl.portal.Secret and gtk for all
other portal implementations.  The configuration file is placed in
/usr/share/xdg-desktop-portal/portals.conf, which has the lowest
priority of all possible locations.  Therefore, it will be ignored
if a desktop environment-specific configuration file is chosen, or if
the user has specified a configuration in /etc/xdg-desktop-portal/.

Fixes: QubesOS/qubes-issues#8839