qvm-copy: support --ignore-symlinks and honor "--"
"--" was stripped from the arguments before being passed to qfile-agent,
so "qvm-copy -- --ignore-symlinks" would ignore symlinks instead of
copying a file named "--ignore-symlinks".  To fix this problem, always
pass "--" to qfile-agent, preceeded by --ignore-symlinks if necessary.
This requires qfile-agent to honor -- as indicating end of options, so
patch it to do that.