Do not use python-daemon library
It was relevant when qubes-firewall was started from SysV init script,
but it isn't the case anymore. Most of its features were disabled
anyway, and redirecting stdin/out/err is handled by systemd anyway.

This is the last usage of the library which isn't available in
Archlinux, and as such qubes-firewall service was broken there. Not
using this library should fix the issue.

Fixes QubesOS/qubes-issues#9238