qubes.Filecopy: support unpacking files with unsafe characters
This is done by means of the --allow-unsafe-characters flag to
qfile-unpacker, and is only done if allow-unsafe-characters is passed as
a service argument.  The service argument is always visible in the
qrexec prompt, so users will know if this flag will be used and can
choose to reject the call.