debian: include all boot-related configs in the package
This especially adds dracut config (that was previously installed in
Fedora only). While dracut is not the default on Debian, it is possible
to use it, so add the config to the package too.

QubesOS/qubes-issues#8649

(cherry picked from commit 8796b2b56e97b81ba3b9184c12c7bce4d8ac448c)