Unify method to query available terminals
Because of some exceptions such as gnome-terminal and kgx, the ordering
was getting convoluted when considering precedence. Instead of adding
new if statements every time a terminal requires precedence, run only
add a new entry to the for loop. Also allows the user to enforce a
terminal with environment variable TERMINAL.