firewall: avoid crash on qubesdb long path errors
Long hostnames result in qubesdb paths that exceed the current 64
character limit.  Instead of crashing the firewall daemon, log the error
and continue without adding the hostname to the DNS records.

This is a workaround, not a fix, for
https://github.com/QubesOS/qubes-issues/9084.  The fix requires either
longer path support in qubesdb or using a different schema so that keys
will not exceed the qubesdb path length limit.