rpm: skip systemctl daemon-reload if systemd is not running
Skip it in chroot or a container to not fail the RPM transaction. While
RPM says it's "non-crtitical error in the %post scriptlet", it does
consider RPM transaction as failed in the end.
Interestingly, that happens only on Fedora 43, earlier versions actually
do treat it as non-critical and consider RPM transaction successful.

Anyway, adjust to not fail the scriptlet and don't worry if the error
should be considered fatal or not.

QubesOS/qubes-issues#10102

(cherry picked from commit 2ac82f0ce9bf3e4440d6e114a7c45525a378b7eb)