qubes-fs-tree-check: Use relative source path for validation
qubes_pure_validate_symbolic_link_v2() checks that the symlink is not to
be written to an absolute path.  qfile-unpacker correctly passes a
relattive path to this function, but qubes-fs-tree-check wrongly passed
an absolute path.  As a result, operations like 'qvm-copy ~/something'
where ~/something contains symlinks are wrongly rejected.

Fixes QubesOS/qubes-issues#10284