Start qubes-bind-dirs.service before systemd-tmpfiles-setup.service
tmpfiles.d may contain configuration to "fix" the permissions on some
directories (like /var/lib/sdwdate on Whonix-Gateway qubes) in the
event an app qube's template is changed and the UIDs/GIDs for some
users in the original template do not match the UIDs/GIDs for the same
users in the new template. This requires that bind mounts from /rw are
set up before systemd-tmpfiles runs, so ensure that the units run in
that order.