qubes-run-terminal: Add support for Alacritty
Alacritty has support for both starting a new session and just opening a new
window. The latter case uses less resources but can only be run if an existing
instance of Alacritty is running.

Add Alacritty as a terminal option in qubes-run-terminal then add a wrapper
script similar to that used by gnome-console to handle the window creation or
first instance.